Ajout de l'ensemble des tests, Jacoco 93%

+++ Alerts/src/main/java/com/safetynet/alerts/service/FirestationService.java
@@ -1,7 +1,6 @@
 package com.safetynet.alerts.service;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 import org.apache.logging.log4j.LogManager;
@@ -12,57 +11,90 @@ import org.springframework.stereotype.Service;
 import com.safetynet.alerts.model.Firestation;
 import com.safetynet.alerts.repository.FirestationRepository;
 
-import lombok.Data;
-
-@Data
 @Service
 public class FirestationService {
-	
+
 	private static Logger logger = LogManager.getLogger(FirestationService.class);
-	
+
 	@Autowired
 	private FirestationRepository firestationRepository;
-	
+
+	// GET a firestation by id
 	public Optional<Firestation> getFirestation(final int id) {
 		return firestationRepository.findById(id);
 	}
-	
-	public List<Firestation> getFirestations() {
-		List<Firestation> firestations = (List<Firestation>) firestationRepository.findAll();
-		logger.info("Response - All mappings:" + firestations);
+
+	// GET all firestations
+	public ArrayList<Firestation> getFirestations() {
+		ArrayList<Firestation> firestations = (ArrayList<Firestation>) firestationRepository.findAll();
 		return firestations;
 	}
-	
+
+	// POST a firestation
 	public Firestation saveFirestation(Firestation firestation) {
 		Firestation savedFirestation = firestationRepository.save(firestation);
-		logger.info("Response - " + savedFirestation + " saved");
 		return savedFirestation;
-	} 	 	 	
-	
+	}
+
+	// PUT a firestation
+	public Firestation updateFirestation(Firestation firestation) {
+		int id = 0;
+		ArrayList<Firestation> firestations = getFirestations();
+		for (Firestation fs : firestations) {
+			if (fs.getAddress().equals(firestation.getAddress())) {
+				id = fs.getId();
+			}
+		}
+		Optional<Firestation> firestationToUpdate = getFirestation(id);
+		if (firestationToUpdate.isPresent()) {
+			Firestation modifiedFirestation = firestationToUpdate.get();
+			// firstName and lastName non modifiable
+			String address = firestation.getAddress();
+			if (address != null) {
+				modifiedFirestation.setAddress(address);
+			}
+			Integer stationNumber = firestation.getStationNumber();
+			if (stationNumber != null) {
+				modifiedFirestation.setStationNumber(stationNumber);
+			}
+			saveFirestation(modifiedFirestation);
+			return modifiedFirestation;
+		} else {
+			return null;
+		}
+	}
+
 	public void deleteFirestationByAddress(final String address) {
 		int id = 0;
-		List<Firestation> firestations = getFirestations();
-		
-		for(Firestation firestation : firestations) {
+		ArrayList<Firestation> firestations = getFirestations();
+		for (Firestation firestation : firestations) {
 			if (firestation.getAddress().equals(address)) {
 				id = firestation.getId();
 			}
 		}
-		firestationRepository.deleteById(id);  
-		logger.info("Response - Mapping for address " + address + " deleted");
-	} 	
-	
+		if (id != 0) {
+			logger.info("Response - Mapping for address " + address + " deleted");
+			firestationRepository.deleteById(id);
+		} else {
+			logger.error("Wrong entry: " + address + " not in database");
+		}
+	}
+
 	public void deleteFirestationByStationNumber(final int stationNumber) {
-		List<Firestation> firestations = getFirestations();
-		List<Integer> stationIds = new ArrayList<Integer>();
-		for(Firestation firestation : firestations) {
+		ArrayList<Firestation> firestations = getFirestations();
+		ArrayList<Integer> stationIds = new ArrayList<Integer>();
+		for (Firestation firestation : firestations) {
 			if (firestation.getStationNumber().equals(stationNumber)) {
 				stationIds.add(firestation.getId());
 			}
 		}
 		for (int id : stationIds) {
-			firestationRepository.deleteById(id);  
+			if (id != 0) {
+				logger.info("Response - Mappings for station number " + stationNumber + " deleted");
+				firestationRepository.deleteById(id);
+			} else {
+				logger.error("Wrong entry: station number " + stationNumber + " not in database");
+			}
 		}
-		logger.info("Response - Mappings for station number " + stationNumber + " deleted");
-	} 
+	}
 }
